renamed forms to be consistent
