+++ playground/mocks/data/idp/idx/authenticator-enroll-ov-sms.json
@@ -30,7 +30,7 @@
         "rel": [
           "create-form"
         ],
-        "name": "enroll-with-another-channel",
+        "name": "select-enrollment-channel",
         "href": "http://localhost:3000/idp/idx/credential/enroll",
         "method": "POST",
         "accepts": "application/vnd.okta.v1+json",
@@ -54,10 +54,6 @@
                     "required": true,
                     "visible": true,
                     "options": [
-                      {
-                        "value": "local",
-                        "label": "Local"
-                      },
                       {
                         "value": "qrcode",
                         "label": "QR code"
