+++ src/v2/ion/RemediationConstants.js
@@ -24,9 +24,8 @@ const FORMS = {
   SELECT_AUTHENTICATOR_ENROLL_DATA: 'select-authenticator-enroll-data',
   AUTHENTICATOR_ENROLLMENT_DATA: 'authenticator-enrollment-data',
   ENROLL_AUTHENTICATOR: 'enroll-authenticator',
-  ENROLL_SELECT_CHANNEL: 'enroll-with-another-channel',
-  ENROLL_WITH_EMAIL: 'enroll-with-email',
-  ENROLL_WITH_SMS: 'enroll-with-sms',
+  SELECT_ENROLLMENT_CHANNEL: 'select-enrollment-channel',
+  ENROLLMENT_CHANNEL_DATA: 'enrollment-channel-data',
   ENROLL_POLL: 'enroll-poll',
   REENROLL_AUTHENTICATOR: 'reenroll-authenticator',
   REENROLL_AUTHENTICATOR_WARNING: 'reenroll-authenticator-warning',
