+++ src/v2/view-builder/utils/LinksUtil.js
@@ -114,12 +114,12 @@ const getSignOutLink = (settings) => {
 };
 
 const getSwitchOVEnrollChannelLink = (appState) => {
-  if (appState.hasRemediationObject('enroll-with-another-channel')) {
+  if (appState.hasRemediationObject(RemediationForms.SELECT_ENROLLMENT_CHANNEL)) {
     return {
       'type': 'link',
       'label': loc('oie.enroll.okta_verify.qrcode.cannotScan', 'login'),
       'name': 'switch-channel',
-      'formName': 'enroll-with-another-channel',
+      'formName': RemediationForms.SELECT_ENROLLMENT_CHANNEL,
     };
   }
 };
